Loaded profile flag
